Move helper to asylum app for better reuse
