+++ project/members/tests/fixtures/notes.py
@@ -1,23 +1,8 @@
 # -*- coding: utf-8 -*-
 import datetime, pytz
 import factory.django, factory.fuzzy
-import loremipsum
-import random
-import re
 from members.models import Member, MemberNote
-
-
-class FuzzyLoremipsum(factory.fuzzy.BaseFuzzyAttribute):
-    fixer = re.compile("[bB]'(.*?)'")
-    fixto = "\g<1>"
-
-    def fuzz(self):
-        ret = ""
-        for outer in range(random.randint(1, 5)):
-            ret += "\n\n"
-            for inner in range(random.randint(3, 10)):
-                ret += self.fixer.sub(self.fixto, loremipsum.get_sentence()).capitalize() + " "
-        return ret.strip()
+from asylum.tests.utils import FuzzyLoremipsum
 
 
 class MemberNoteFactory(factory.django.DjangoModelFactory):
